pyre-jit-trace: extract vstack_mirror helpers to vstack_mirror.rs
Move the value-stack mirror (classify/reconcile/reseed/seed/step/
enter-handler) and the pc-map block-head helpers it steps through into
vstack_mirror.rs.
